Unclexo

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,18 @@ use App\Models\Post;
 
 class PostController extends Controller
 {
+    public function index()
+    {
+        $posts = auth()->user()->posts;
+
+        return view('posts.index', compact('posts'));
+    }
+
+    public function show(Post $post)
+    {
+        return view('posts.show', compact('post'));
+    }
+
     public function create()
     {
         return view('posts.create');
@@ -19,13 +31,15 @@ class PostController extends Controller
         return redirect($post->path());
     }
 
-    public function show(Post $post)
+    public function edit(Post $post)
     {
-        return view('posts.show', compact('post'));
+        return view('posts.edit', compact('post'));
     }
 
     public function update(PostRequest $request, Post $post)
     {
+        $this->authorize('update', $post); // Check Gate authorization api
+
         $post->update($request->validated());
 
         return redirect($post->path());
@@ -33,6 +47,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        $this->authorize('update', $post);
+
         $post->delete();
 
         return redirect()->route('posts.index');
